[STEP05] 김성인 - e-commerce

+++ src/main/java/kr/hhplus/be/server/domain/user/dto/BalanceChargeRequest.java
@@ -2,4 +2,6 @@ package kr.hhplus.be.server.domain.user.dto;
 
 import jakarta.validation.constraints.Min;
 
-public record BalanceChargeRequest(@Min(value=1, message = "amount must be over zero") int amount) {}
+import java.math.BigDecimal;
+
+public record BalanceChargeRequest(@Min(value=1, message = "amount must be over zero") BigDecimal amount) {}
